Complete attorney screen modal and submission
