+++ js/screens/onboarding/AttorneyScreen.js
@@ -6,40 +6,27 @@ import {
   StyleSheet,
   TouchableWithoutFeedback,
   Keyboard,
-  Text,
-  TouchableOpacity,
 } from 'react-native';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
 import { useTranslation } from 'react-i18next';
 import routes from '../../navigation/routes';
-import { textStyles, colors } from '../../styles';
 import OnboardingTitle from '../../components/OnboardingTitle';
 import OnboardingButtons from '../../components/OnboardingButtons';
 import AttorneyForm from '../../components/AttorneyForm';
-import AttorneyModal from '../../components/AttorneyModal';
 
 const onBoardingRoutes = routes.onboarding;
 
 const AttorneyScreen = ({ navigation }) => {
   const { t } = useTranslation();
-  const [modalVisible, setModalVisible] = useState(false);
-  const [defaultUsed, setDefaultUsed] = useState(false);
 
-  const [name, setName] = useState('');
-  const [phoneNumber, setPhoneNumber] = useState('');
-
-  const onModalSubmit = () => {
-    setDefaultUsed(true);
-    setModalVisible(false);
+  const [shouldFormSubmit, setShouldFormSubmit] = useState(false);
+  const onSubmit = () => {
+    setShouldFormSubmit(true);
+    navigation.navigate(onBoardingRoutes.complete);
   };
 
   return (
     <KeyboardAwareScrollView contentContainerStyle={styles.container}>
-      <AttorneyModal
-        isVisible={modalVisible}
-        setIsVisible={setModalVisible}
-        onSubmit={onModalSubmit}
-      />
       <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
         <View style={styles.inner}>
           <View style={styles.contentContainer}>
@@ -47,27 +34,12 @@ const AttorneyScreen = ({ navigation }) => {
               title={t('select_attorney')}
               subtitle={t('select_attorney_subtitle')}
             />
-            <AttorneyForm
-              name={name}
-              setName={setName}
-              phoneNumber={phoneNumber}
-              setPhoneNumber={setPhoneNumber}
-              isLocked={defaultUsed}
-            />
-            <TouchableOpacity
-              style={styles.noAttorney}
-              onPress={() => setModalVisible(true)}
-            >
-              <Text style={[textStyles.h3, { color: colors.primary }]}>
-                {t('no_attorney')}
-              </Text>
-            </TouchableOpacity>
-            <Text>{defaultUsed ? 'true' : 'false'}</Text>
+            <AttorneyForm shouldFormSubmit={shouldFormSubmit} />
           </View>
           <View style={styles.buttonContainer}>
             <OnboardingButtons
               onRightPress={() => navigation.pop()}
-              onLeftPress={() => navigation.navigate(onBoardingRoutes.complete)}
+              onLeftPress={onSubmit}
               rightTitle={t('back')}
               leftTitle={t('next')}
               nextIsDisabled={false}
@@ -109,9 +81,5 @@ const styles = StyleSheet.create({
     width: '95%',
     flexDirection: 'row',
     justifyContent: 'space-between',
-  },
-  noAttorney: {
-    alignItems: 'flex-end',
-    paddingRight: 10,
   },
 });
